fix(form): fix resetting dependent fields during schema changes in lists
