+++ src/components/form/form.tsx
@@ -18,6 +18,7 @@ import {
     ObjectFieldTemplate,
 } from './templates';
 import { SchemaField as CustomSchemaField } from './fields/schema-field';
+import { ArrayField as CustomArrayField } from './fields/array-field';
 import { widgets } from './widgets';
 import { createRandomString } from '../../util/random-string';
 import Ajv from 'ajv';
@@ -131,6 +132,7 @@ export class Form {
                     },
                     fields: {
                         SchemaField: CustomSchemaField,
+                        ArrayField: CustomArrayField,
                     },
                 },
                 []
